+++ calendar-app/src/utils/updateTimeframe.ts
@@ -1,4 +1,3 @@
-import { clearEvents } from "../api/displayEvents";
 function timeframeUpdate(currentDate: Date) {
   //clearEvents();
   //displayEvents(currentDate);
